tx: move files into dirs
