+++ packages/tx/src/index.ts
@@ -1,8 +1,8 @@
-export { FeeMarketEIP1559Transaction } from './eip1559Transaction.js'
-export { AccessListEIP2930Transaction } from './eip2930Transaction.js'
-export { BlobEIP4844Transaction } from './eip4844Transaction.js'
-export { EOACodeEIP7702Transaction } from './eip7702Transaction.js'
-export { LegacyTransaction } from './legacyTransaction.js'
+export { FeeMarketEIP1559Transaction } from './1559/tx.js'
+export { AccessListEIP2930Transaction } from './2930/tx.js'
+export { BlobEIP4844Transaction } from './4844/tx.js'
+export { EOACodeEIP7702Transaction } from './7702/tx.js'
+export { LegacyTransaction } from './legacy/tx.js'
 export {
   createTxFromBlockBodyData,
   createTxFromJsonRpcProvider,
