+++ packages/tx/src/legacy/tx.ts
@@ -12,16 +12,16 @@ import {
 } from '@ethereumjs/util'
 import { keccak256 } from 'ethereum-cryptography/keccak.js'
 
-import { BaseTransaction } from './baseTransaction.js'
-import * as Legacy from './capabilities/legacy.js'
-import { Capability, TransactionType } from './types.js'
+import { BaseTransaction } from '../baseTransaction.js'
+import * as Legacy from '../capabilities/legacy.js'
+import { Capability, TransactionType } from '../types.js'
 
 import type {
   TxData as AllTypesTxData,
   TxValuesArray as AllTypesTxValuesArray,
   JsonTx,
   TxOptions,
-} from './types.js'
+} from '../types.js'
 import type { Common } from '@ethereumjs/common'
 
 type TxData = AllTypesTxData[TransactionType.Legacy]
